Downloads/project4/requirements.txt
#detection
matplotlib
numpy
pandas
seaborn
scikit-learn
tensorflow

#types
keras
opencv-python
pillow
scikit-learn